refactor: standardize use case class names and update import paths
