--- src/main/java/com/gresk/modules/account/application/port/in/LoginUseCase.java
@@ -1,9 +1,0 @@
-package com.gresk.modules.account.application.port.in;
-
-import com.gresk.infrastructure.port.AuthToken;
-import com.gresk.modules.account.application.command.LoginCommand;
-
-public interface LoginUseCase {
-
-    AuthToken execute(LoginCommand command);
-}
