--- src/main/java/com/gresk/modules/account/application/port/in/RegisterUserAccountUseCase.java
@@ -1,9 +1,0 @@
-package com.gresk.modules.account.application.port.in;
-
-import com.gresk.modules.account.application.command.RegisterUserAccountCommand;
-import com.gresk.modules.account.domain.model.AccountId;
-
-public interface RegisterUserAccountUseCase {
-
-    AccountId execute(RegisterUserAccountCommand command);
-}
